+++ activestorage/CHANGELOG.md
@@ -1,2 +1,7 @@
+*   `ActiveStorage::Blob#open` can now be used without passing a block. When using this form the returned temporary file
+    must be unlinked manually.
+
+    *Bart de Water*
+
 
 Please check [7-2-stable](https://github.com/rails/rails/blob/7-2-stable/activestorage/CHANGELOG.md) for previous changes.